vars: correctly compare EFIVARFS constant
On 32-bit linux st.Type may be int32, whereas the magic constant in
golang is uint32. In C, it doesn't matter as int32/uint32 constant are
compared as memory rather than as values with range checking. But in
Go, it matters.

Imho Golang sys unix bindings should probably fix this, and have the
St.Type magic constants be processed as signed values, as needed on
per-arch basis.

Until that's fixed in Golang, redeclare EFIVARFS magic constant with
correct sign.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>