vars: Unlink the path when writing an empty payload
EFI variables are deleted by writing an empty payload to them, which
must be a signed empty payload for authenticated variables. However,
the kernel does some validation of the variable payload when it is
set, and this validation is too strict for Boot* variables (it rejects
an empty payload).

efivarfs will set a variable to an empty payload when a path is unlinked,
so use this when trying to set a variable payload to zero bytes.