Use CertLikelyTrustAnchor instead of IsRootCert and refactor how it works
IsRootCert is a bit strange because it builds candidate certificate
chains by verifying that certificate signatures are valid, but it
performs no other verification - it doesn't verify the actual signature
and doesn't check certificate usage etc.

Rename it to CertLikelyTrustAnchor instead to be clearer about the
guarantees of the API, and also stop relying on valid signatures when
building candidate certificate chains - instead, use the certificates'
subject, issuer, SKID, AKID, public algorithm and signature algorithm
fields to determine whether one certificate is likely an issuer of
another certificate.