wincert: Relax CertLikelyTrustAnchor when signed by the trust anchor
[WinCertificatePKCS7,WinCertificateAuthenticode].CertLikelyTrustAnchor
is meant to return true if the supplied trust anchor actually created
the signature, and it should do this regardless of whether the trust
anchor is self-signed or not. In order to handle this case, it compares
the raw DER encoded certificate blobs.

However, we abuse this API a little bit in secboot because certificate
path building doesn't perform any cryptographic operations and doesn't
really require the entire contents of the trust anchor - it only needs
the fields required for building a path (subject, SKID and public key
algorithm). Only supplying these fields doesn't work in the case where
the trust anchor created the signature because the raw DER encoded
certificates are compared.

Relax this a little bit and assume that trust anchor is the signer if
they have the same subject, issuer, SKID, AKID, signature algorithm and
public key algorithm. Also, only perform this check (trust anchor ==
signer) if no other chains were built successfully.