Pass a context.Context to functions that interact with EFI variables
A previous commit provided for a way to mock the VarsBackend, but this
is only really suitable in test code because it doesn't provide a way
to mock the backend in a way that is specific to a go routine - it is
global to an entire process.

Secboot already has a way of abstracting variable reads (via a
HostEnvironment type) but it is using an increasing number of functions
provided by this package that perform variable reads indirectly. It
isn't really feasible to mock every single one of these, or for secboot
to mock the global VarsBackend in production code, given that there may be
more than one go routine performing operations on EFI variables at the
same time.

The best idea I could come up with here was to pass a context.Context to
every function that interacts with EFI variables and have a VarsBackend
keyed from here. There is a default DefaultVarContext which is initialized
to use the system default mechanism of interacting with EFI variables, but
it provides a way for other packages to mock the VarsBackend by
providing their own context to calls with the VarsBackendKey on it.