wincert: Add X509CertID interface and corresponding APIs
This package doesn't need full X.509 certificates in order to determine
if one certificate likely issued another, yet the public APIs accepted a
x509.Certificate.

To avoid embedding full certificate data in secboot, we embed the
information required to determine whether a certificate is likely the
trust anchor of a signature and then construct a partial
x509.Certificate which is then passed to this package.